payment_handler.py
from telegram import LabeledPrice, Update, InlineKeyboardButton, InlineKeyboardMarkup
from telegram.ext import ContextTypes
from telegram.constants import ParseMode
from typing import Dict, List, Optional, Any
import logging
from database import Database
from config import Config


class PaymentHandler:
    def __init__(self, db: Database):
        self.db = db
        self.config = Config()
        self.logger = logging.getLogger(__name__)
        
        # TON wallet address for direct cryptocurrency donations
        self.ton_wallet = "UQD_your_ton_wallet_address_here"  # Replace with actual wallet
        
        # Telegram payment provider token (would be set in production)
        self.payment_provider_token = None  # Set this in production

    async def process_stars_donation(self, update: Update, context: ContextTypes.DEFAULT_TYPE, amount: int) -> None:
        """Process a Stars donation."""
        user_id = update.effective_user.id
        
        # In a real implementation, this would integrate with Telegram's Stars API
        # For now, we'll just simulate the process
        
        # Create a thank you message
        message = f"""
⭐ Thank you for your {amount} Stars donation!

Your support helps keep the bot running and improving.

Note: Stars donation processing will be implemented when the feature is fully available in the Telegram Bot API.
        """
        
        # Create a keyboard with a thank you button
        keyboard = InlineKeyboardMarkup([
            [InlineKeyboardButton("❤️ You're Welcome!", callback_data="donate_thanks")],
            [InlineKeyboardButton("⬅️ Back to Donation Options", callback_data="donate_main")]
        ])
        
        # Send the message
        if update.callback_query:
            await update.callback_query.edit_message_text(
                message,
                reply_markup=keyboard,
                parse_mode=ParseMode.HTML
            )
        else:
            await update.message.reply_text(
                message,
                reply_markup=keyboard,
                parse_mode=ParseMode.HTML
            )
        
        # Log the donation
        await self.db.log_donation(user_id, amount, "STARS", "telegram_stars")
        
        self.logger.info(f"User {user_id} donated {amount} Stars")

    async def process_ton_donation(self, update: Update, context: ContextTypes.DEFAULT_TYPE, amount: float) -> None:
        """Process a TON donation."""
        user_id = update.effective_user.id
        
        # Create a message with the TON wallet address
        message = f"""
💎 <b>TON Donation: {amount} TON</b>

Please send {amount} TON to the following wallet address:

<code>{self.ton_wallet}</code>

After sending, please contact @YourSupportUsername with the transaction hash to confirm your donation.

Thank you for your support! ❤️
        """
        
        # Create a keyboard with a thank you button
        keyboard = InlineKeyboardMarkup([
            [InlineKeyboardButton("📋 Copy Wallet Address", callback_data="donate_copy_wallet")],
            [InlineKeyboardButton("⬅️ Back to Donation Options", callback_data="donate_main")]
        ])
        
        # Send the message
        if update.callback_query:
            await update.callback_query.edit_message_text(
                message,
                reply_markup=keyboard,
                parse_mode=ParseMode.HTML
            )
        else:
            await update.message.reply_text(
                message,
                reply_markup=keyboard,
                parse_mode=ParseMode.HTML
            )
        
        # Log the donation intent
        await self.db.log_donation(user_id, amount, "TON", "ton_wallet")
        
        self.logger.info(f"User {user_id} initiated a {amount} TON donation")

    async def handle_custom_stars_amount(self, update: Update, context: ContextTypes.DEFAULT_TYPE) -> None:
        """Handle custom Stars amount input."""
        # This would be implemented when Telegram's Stars API is available
        message = """
⭐ <b>Custom Stars Donation</b>

Please enter the number of Stars you would like to donate:

<i>Note: Custom Stars donation will be implemented when the feature is fully available in the Telegram Bot API.</i>
        """
        
        keyboard = InlineKeyboardMarkup([
            [
                InlineKeyboardButton("⭐ 10", callback_data="donate_stars_10"),
                InlineKeyboardButton("⭐ 25", callback_data="donate_stars_25"),
                InlineKeyboardButton("⭐ 50", callback_data="donate_stars_50")
            ],
            [InlineKeyboardButton("⬅️ Back", callback_data="donate_stars")]
        ])
        
        await update.callback_query.edit_message_text(
            message,
            reply_markup=keyboard,
            parse_mode=ParseMode.HTML
        )

    async def handle_custom_ton_amount(self, update: Update, context: ContextTypes.DEFAULT_TYPE) -> None:
        """Handle custom TON amount input."""
        message = """
💎 <b>Custom TON Donation</b>

Please select a TON amount or enter a custom value:
        """
        
        keyboard = InlineKeyboardMarkup([
            [
                InlineKeyboardButton("💎 0.1", callback_data="donate_ton_0.1"),
                InlineKeyboardButton("💎 0.5", callback_data="donate_ton_0.5"),
                InlineKeyboardButton("💎 1", callback_data="donate_ton_1")
            ],
            [InlineKeyboardButton("⬅️ Back", callback_data="donate_ton")]
        ])
        
        await update.callback_query.edit_message_text(
            message,
            reply_markup=keyboard,
            parse_mode=ParseMode.HTML
        )

    async def handle_payment_success(self, update: Update, context: ContextTypes.DEFAULT_TYPE) -> None:
        """Handle successful payment."""
        user_id = update.effective_user.id
        payment = update.message.successful_payment
        
        # Log the payment
        await self.db.log_donation(
            user_id, 
            payment.total_amount / 100,  # Convert from cents
            payment.currency,
            "telegram_payment"
        )
        
        # Send thank you message
        await update.message.reply_text(
            f"Thank you for your donation of {payment.total_amount / 100} {payment.currency}! ❤️\n\n"
            f"Your support helps keep the bot running and improving."
        )
        
        self.logger.info(f"User {user_id} made a successful payment of {payment.total_amount / 100} {payment.currency}")

    async def get_donation_stats(self) -> Dict:
        """Get donation statistics."""
        total_donations = await self.db.get_total_donations()
        
        # Format the stats
        stats = {
            "total_by_currency": total_donations,
            "donor_count": await self._count_unique_donors()
        }
        
        return stats

    async def _count_unique_donors(self) -> int:
        """Count unique donors."""
        # This would be implemented in the database class
        # For now, return a placeholder
        return 0
